# Improve monitor performance