--- a/src/api/app/views/webui2/webui/project/monitor.html.haml
+++ b/src/api/app/views/webui2/webui/project/monitor.html.haml
@@ -20,7 +20,7 @@
             %tr
               %th
               - @repoarray.each do |repo, archlist|
-                - archlist.sort.each do |arch|
+                - archlist.sort!.each do |arch|
                   %th.text-center
                     = webui2_repository_status_icon(status: @repostatushash[repo][arch], html_class: 'fa-xs mr-1')
                     = arch
@@ -30,7 +30,7 @@
                 %td
                   = link_to word_break(packname, 40), controller: :package, action: :show, package: packname, project: @project.to_s
                 - @repoarray.each do |repo, archlist|
-                  - archlist.sort.each do |arch|
+                  - archlist.each do |arch|
                     %td.text-center
                       = webui2_arch_repo_table_cell(repo, arch, packname, nil, false)
 :javascript