This is improving both the old UI and the new one (Bootstrap), although it has a much bigger impact in the new one.

Solves https://github.com/openSUSE/open-build-service/issues/6554

Times with 74 packages and 40 repositories:

# NEW UI

### Before

14151ms (Views: 12719.2ms | ActiveRecord: 191.2ms | Backend: 726.9ms)

### Now

3750ms (Views: 2785.9ms | ActiveRecord: 19.9ms | Backend: 724.4ms)

# OLD UI

### Before

4685ms (Views: 3351.1ms | ActiveRecord: 176.3ms | Backend: 680.2ms)

### Now

3653ms (Views: 2681.3ms | ActiveRecord: 21.3ms | Backend: 715.8ms)