@Ana06 2019-01-25
![](https://github.trello.services/images/mini-trello-icon.png) [Improve Monitor Page Performance](https://trello.com/c/Pj3pWPd3/1036-improve-monitor-page-performance)