@Ana06 2019-01-26
@coolo 

> Your benchmark looks like cheating as you didn't change the backend calls, but picked runtimes where the backend changes dramatically :)

I put the benchmark just to show that this is not solving the other issue and that we go significatively down in SQL. It is anyway not accurate because depends on if there are other things running.