@coolo 2019-01-29
Referencing this:
> But I wonder: app/mixins/webui/load_buildresults.rb and spec/controllers/webui/project_controller_spec.rb
> access this variable. I guess this is obsolete